+++ conference-manager/ms-conference-api/src/exceptions/BadRequest.exception.ts
