+++ conference-manager/ms-conference-api/src/exceptions/BadRequest.ts
@@ -1,7 +1,7 @@
 import { Catch, HttpException, HttpStatus } from '@nestjs/common';
 
 @Catch(HttpException)
-export class ChupitosBadRequestException extends HttpException {
+export class BadRequestException extends HttpException {
   constructor(message: string) {
     super(`Bad request exception: ${message}`, HttpStatus.BAD_REQUEST)
   }
