+++ conference-manager/ms-conference-api/src/exceptions/NotFound.exception.ts
@@ -1,7 +1,7 @@
 import { Catch, HttpException, HttpStatus } from '@nestjs/common'
 
 @Catch(HttpException)
-export class ChupitosNotFoundException extends HttpException {
+export class NotFoundException extends HttpException {
   constructor(message: string) {
     super(`Not found exception: ${message}`, HttpStatus.NOT_FOUND)
   }
