+++ conference-manager/ms-conference-api/src/modules/headquarter/headquarter.service.ts
@@ -4,7 +4,7 @@ import { InjectModel } from '@nestjs/mongoose'
 import { Injectable, Logger } from '@nestjs/common'
 import { HeadquarterResponse } from './interfaces/headquarter-response'
 import HeadquarterMapper from './headquarter.mapper'
-import { ChupitosNotFoundException } from '../../exceptions/chupitos-not-found.exception'
+import { NotFoundException } from '../../exceptions/NotFound.exception'
 
 @Injectable()
 export class HeadquarterService {
@@ -25,7 +25,7 @@ export class HeadquarterService {
     const headquarterById = await this.headquarterModel.findById(headquarterId)
 
     if (!headquarterById)
-      throw new ChupitosNotFoundException(
+      throw new NotFoundException(
         `Error at get headquarter by id service, _id: ${headquarterById} was not found`
       )
     return headquarterById
